python-pom/requirements.txt
Appium-Python-Client>=0.28
